requirements.txt
ansible==9.5.1
ansible-core==2.16.6
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-compute==1.18.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyaml==24.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyYAML==6.0.1
requests==2.31.0
resolvelib==1.0.1
rsa==4.9
terrascript==0.9.0
urllib3==2.2.1
